# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "API for the Reward Loyalty script.",
        "contact": {
            "email": "support@nowsquare.com"
        },
        "version": "1.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://reward-loyalty-demo.nowsquare.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/partner/cards": {
            "get": {
                "tags": [
                    "Partner",
                    "Card"
                ],
                "summary": "Fetch all active cards associated with a partner",
                "description": "Endpoint to retrieve all active cards linked to the authenticated partner, where the associated club is also active",
                "operationId": "getPartnerCards",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cards retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Card"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/partner/cards/{cardId}": {
            "get": {
                "tags": [
                    "Partner",
                    "Card"
                ],
                "summary": "Fetch an active card using its ID",
                "description": "Endpoint for a partner to retrieve a specific active card that they own by its ID",
                "operationId": "getCard",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardId",
                        "in": "path",
                        "description": "Unique identifier of the card to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Card"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "Not found."
                    }
                },
                "type": "object"
            },
...
            "Card": {
                "required": [
                    "id",
                    "club_id",
                    "name",
                    "head",
                    "title",
                    "description",
                    "unique_identifier",
                    "issue_date",
                    "expiration_date",
                    "bg_color",
                    "bg_color_opacity",
                    "text_color",
                    "text_label_color",
                    "qr_color_light",
                    "qr_color_dark",
                    "currency",
                    "initial_bonus_points",
                    "points_expiration_months",
                    "currency_unit_amount",
                    "points_per_currency",
                    "min_points_per_purchase",
                    "max_points_per_purchase",
                    "is_visible_by_default",
                    "is_visible_when_logged_in",
                    "total_amount_purchased",
                    "number_of_points_issued",
                    "last_points_issued_at",
                    "number_of_points_redeemed",
                    "number_of_rewards_redeemed",
                    "last_reward_redeemed_at",
                    "views",
                    "last_view",
                    "created_by",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the Card",
                        "type": "string",
                        "example": "50510829749146291"
                    },
                    "club_id": {
                        "description": "Unique identifier for the Club",
                        "type": "string",
                        "example": "51209532520667617"
                    },
                    "name": {
                        "description": "Name of the Card",
                        "type": "string"
                    },
                    "head": {
                        "description": "Multilingual card name",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Gesunde Ernährung"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Healthy Eats"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "Comida Saludable"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Repas Sains"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Gezond Eten"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Comidas Saudáveis"
                            }
                        },
                        "type": "object"
                    },
                    "title": {
                        "description": "Multilingual card title",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Salat Ersparnisse"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Salad Savings"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "Ahorros en Ensaladas"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Économies sur les Salades"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Salade Besparingen"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Economia em Salada"
                            }
                        },
                        "type": "object"
                    },
                    "description": {
                        "description": "Multilingual card description",
                        "properties": {
                            "de_DE": {
                                "type": "string",
                                "example": "Belohnung für gesunde Entscheidungen!"
                            },
                            "en_US": {
                                "type": "string",
                                "example": "Get rewarded for healthy choices!"
                            },
                            "es_ES": {
                                "type": "string",
                                "example": "¡Obtén recompensas por elecciones saludables!"
                            },
                            "fr_FR": {
                                "type": "string",
                                "example": "Soignez vos choix et soyez récompensé(e) !"
                            },
                            "nl_NL": {
                                "type": "string",
                                "example": "Word beloond voor gezonde keuzes!"
                            },
                            "pt_BR": {
                                "type": "string",
                                "example": "Seja recompensado por escolhas saudáveis!"
                            }
                        },
                        "type": "object"
                    },
                    "unique_identifier": {
                        "description": "Unique identifier of the Card",
                        "type": "string"
                    },
                    "issue_date": {
                        "description": "Issue date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiration_date": {
                        "description": "Expiration date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "bg_color": {
                        "description": "Background color of the Card",
                        "type": "string",
                        "example": "#CCCCCC"
                    },
                    "bg_color_opacity": {
                        "description": "Background color opacity of the Card",
                        "type": "integer",
                        "format": "int32",
                        "example": 80
                    },
                    "text_color": {
                        "description": "Text color of the Card",
                        "type": "string",
                        "example": "#333333"
                    },
                    "text_label_color": {
                        "description": "Text label color of the Card",
                        "type": "string",
                        "example": "#333333"
                    },
                    "qr_color_light": {
                        "description": "QR code light color of the Card",
                        "type": "string",
                        "example": "#CCCCCC"
                    },
                    "qr_color_dark": {
                        "description": "QR code dark color of the Card",
                        "type": "string",
                        "example": "#222222"
                    },
                    "currency": {
                        "description": "Currency of the Card",
                        "type": "string",
                        "example": "USD"
                    },
                    "initial_bonus_points": {
                        "description": "Initial bonus points of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "points_expiration_months": {
                        "description": "Points expiration months of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "currency_unit_amount": {
                        "description": "Currency unit amount of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "points_per_currency": {
                        "description": "Points per currency of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "min_points_per_purchase": {
                        "description": "Minimum points per purchase of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "max_points_per_purchase": {
                        "description": "Maximum points per purchase of the Card",
                        "type": "integer",
                        "format": "int64"
                    },
                    "is_visible_by_default": {
                        "description": "Status of the Card if it's visible by default",
                        "type": "integer",
                        "format": "int32"
                    },
                    "is_visible_when_logged_in": {
                        "description": "Status of the Card if it's visible when logged in",
                        "type": "integer",
                        "format": "int32"
                    },
                    "total_amount_purchased": {
                        "description": "Total amount purchased of the Card",
                        "type": "integer",
                        "format": "int64"
                    },
                    "number_of_points_issued": {
                        "description": "Number of points issued for the Card",
                        "type": "integer",
                        "format": "int64"
                    },
                    "last_points_issued_at": {
                        "description": "Last time points were issued for the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "number_of_points_redeemed": {
                        "description": "Number of points redeemed from the Card",
                        "type": "integer",
                        "format": "int64"
                    },
                    "number_of_rewards_redeemed": {
                        "description": "Number of rewards redeemed from the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "last_reward_redeemed_at": {
                        "description": "Last time a reward was redeemed from the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "views": {
                        "description": "Number of views of the Card",
                        "type": "integer",
                        "format": "int32"
                    },
                    "last_view": {
                        "description": "Last view of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "meta": {
                        "description": "Meta of the Card",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update date of the Card",
                        "type": "string",
                        "format": "date-time"
                    },
                    "balance": {
                        "description": "The member's balance. This field has a value of 0 or higher for API calls where a member is authenticated. If the balance is not applicable or unavailable, its value is -1.",
                        "type": "integer",
                        "format": "int32",
                        "example": -1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "partner_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    }
}